moved redux and clean
