+++ src/components/redux/reducers/content.js
